Refactored EditableInput
